Refactor DiContainer and DiInfo classes
This commit refactors the DiContainer and DiInfo classes by adding a new static method to the DiContainer class called Get. This method allows retrieving the container associated with a given instance. Additionally, the DiInfo class now has a new static method called GetContainer, which returns the container associated with a given instance. These changes improve the functionality and usability of the classes.